requirements.txt
torch>=2.3
transformers>=4.45
peft>=0.13
accelerate
pyyaml
